- GithubRepositoryRepository interfaceの作成 - modelの移動
