+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/RepositorySearchViewModel.kt
@@ -3,7 +3,6 @@
  */
 package jp.co.yumemi.android.code_check
 
-import android.os.Parcelable
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -11,13 +10,12 @@ import io.ktor.client.request.get
 import io.ktor.client.request.header
 import io.ktor.client.request.parameter
 import jp.co.yumemi.android.code_check.MainActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.code_check.core.data.model.GithubRepository
+import jp.co.yumemi.android.code_check.core.data.model.RepositorySearchResponse
 import jp.co.yumemi.android.code_check.network.HttpClientSingleton.client
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import kotlinx.parcelize.Parcelize
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerializationException
 import java.net.UnknownHostException
 import java.util.Date
@@ -34,7 +32,7 @@ class RepositorySearchViewModel : ViewModel() {
     val errorState = _errorState.asStateFlow()
 
     // StateFlowを使用して検索結果を保持
-    private val _searchResults = MutableStateFlow<List<RepositoryInfoItem>>(emptyList())
+    private val _searchResults = MutableStateFlow<List<GithubRepository>>(emptyList())
     val searchResults = _searchResults.asStateFlow()
 
     /**
@@ -91,43 +89,6 @@ class RepositorySearchViewModel : ViewModel() {
     }
 }
 
-
-@Parcelize
-@Serializable
-data class RepositorySearchResponse(
-    val items: List<RepositoryInfoItem>
-) : Parcelable
-
-
-/**
- * Githubのリポジトリ情報
- */
-@Parcelize
-@Serializable
-data class RepositoryInfoItem(
-    @SerialName("full_name")
-    val name: String,
-    @SerialName("owner")
-    val owner: Owner,
-    @SerialName("language")
-    val language: String?,
-    @SerialName("stargazers_count")
-    val stargazersCount: Long,
-    @SerialName("watchers_count")
-    val watchersCount: Long,
-    @SerialName("forks_count")
-    val forksCount: Long,
-    @SerialName("open_issues_count")
-    val openIssuesCount: Long,
-) : Parcelable
-
-
-@Parcelize
-@Serializable
-data class Owner(
-    @SerialName("avatar_url")
-    val avatarUrl: String,
-) : Parcelable
 /**
  * エラーの状態を表すsealed interface
  */
