+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/core/data/GithubRepositoryRepository.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.code_check.core.data
+
+import jp.co.yumemi.android.code_check.core.data.model.GithubRepository
+
+interface GithubRepositoryRepository {
+    suspend fun getGithubRepositoryList(): List<GithubRepository>
+}
