+++ app/src/main/res/navigation/nav_graph.xml
@@ -21,8 +21,8 @@
         android:label="@string/repository_info"
         tools:layout="@layout/fragment_repository_info">
         <argument
-            android:name="repositoryInfoItem"
-            app:argType="jp.co.yumemi.android.code_check.RepositoryInfoItem" />
+            android:name="githubRepository"
+            app:argType="jp.co.yumemi.android.code_check.core.data.model.GithubRepository" />
     </fragment>
 
 </navigation>
